requirements.txt
pandas
openpyxl
python-docx
pdfplumber
PyMuPDF
requests
tqdm
pytest
pytest-mock
python-dotenv